requirements.txt
PyYAML
Jinja2
jsonschema
